packages/python/examples/tencent/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
requests==2.31.0
sseclient-py==1.7.2
urllib3==2.0.5
dynaconf
tcvectordb